+++ wakamiti-api/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Path temporal placeholder for a specific instant.
 - New methods in JsonUtils to convert to any type.
 - Abstract HttpClient.
+- StringUtils class.
 
 ### Changed
 - Authentication utility classes are moved to the api library.
@@ -22,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - `ID([\w-]+)` is now the default value of `idTagPattern`.
 - The map util accepts null values.
+- Resource relative path.
 
 
 ## [2.4.1] - 2024-04-16